+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.0 2019-11-20
+### Added
+- Improved filtering of items in Mention plugin.
+- Mention list can now be hidden by pressing escape.
+- `maxItems` prop added to Mention if only x number of items should show at one time.
+
 ## 0.5.2 2019-11-08
 ### Fixes
 - Export props for TS